Allow emitting `Element` values
